% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessModel.R
\name{assessModel}
\alias{assessModel}
\title{Assess the training model results}
\usage{
assessModel(object, obs)
}
\arguments{
\item{object}{the object of the stack_ensemble model}

\item{obs}{the original values in test dataset}
}
\value{
a list with components:
the prediction of base ensemble model
the prediction of stack ensemble model if it is a stack ensemble object
the assessment for base ensemble model
the assessment for stack ensemble model if it is a stack ensemble object
}
\description{
Assess the training models
}
\examples{
\dontrun{
#Spliting training set into two parts based on outcome: 80\% and 20\%
set.seed(1234)
data(envir_example)
index <-  caret::createDataPartition(envir_example$PM2.5, p=0.8, list=FALSE)
trainSet <- envir_example[ index,]
testSet <- envir_example[-index,]

#set the predictor "PM2.5" and dependence variables
predictor <- colnames(envir_example)[-c(1,6)]
dependence <- c("PM2.5")

##ensemble 6 kinds of multi models
sl_lib <- c("SL.xgboost","SL.randomForest","SL.ranger")
#get the predict model
ensemble_model <- predictModel(Y = trainSet[,dependence],X = trainSet[,predictor],
newX = testSet[,predictor], base_model = sl_lib)

model_assess <- assessModel(object = ensemble_model,obs = testSet$PM2.5)
}

}
