% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.predictModel_parallel.R
\name{CV.predictModel_parallel}
\alias{CV.predictModel_parallel}
\title{The external cross validation analysis for SL model with paralleling computing}
\usage{
CV.predictModel_parallel(
  Y,
  X,
  base_model,
  number_cores,
  method = "method.NNLS",
  cvControl = list(),
  ...
)
}
\arguments{
\item{Y}{the dependence, like certain environmental pollutants(eg. PM2.5) in training data that we
want to predict}

\item{X}{the other predictors(eg. temp, wind speed) in training data that we
want to use to predict air pollutants}

\item{base_model}{character string in format of "SL.xgboost","SL.randomForest"}

\item{number_cores}{the number of CPU cores used in parallel computing}

\item{method}{the method of ensemble the base models, "method.NNLS" (the default)}

\item{cvControl}{the list to control the cross validation,default with 10-fold random CV}

\item{...}{other paramaters that belong to 'SuperLearner' pacakge}
}
\value{
a list including
the cross validation model training information;
the cross validation result
}
\description{
make the external cross-validation analysis for SL model
}
\examples{
\dontrun{
# Let's predict the PM2.5 values with machine learming methods

# Splitting training set into two parts based on outcome: 80\% and 20\%
set.seed(1234)
data("envir_example")
size <-
  caret::createDataPartition(y = envir_example$PM2.5, p = 0.8, list = FALSE)
trainset <- envir_example[size, ]
testset <- envir_example[-size, ]

# set the predictor "PM2.5" and dependence variables
y <- c("PM2.5")
x <- colnames(envir_example[-c(1, 6)]) # except "date" and "PM2.5"

pred_m1_cv <- CV.predictModel_parallel(
  Y = trainset[, y],
  X = trainset[, x],
  base_model = c("SL.xgboost", "SL.ranger"),
  number_cores = 4
)
}

}
