% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.R
\name{plot_point}
\alias{plot_point}
\title{To create the scatter plot with the regression line}
\usage{
plot_point(pre, obs, ...)
}
\arguments{
\item{pre}{the prediction value}

\item{obs}{the original value}

\item{...}{the other parameters of package ggplot2}
}
\value{
the scatter plot with the regression formula
}
\description{
check the relationship between the prediction and observation
}
\examples{

plot_point(pre = sample(1:100,100,replace = TRUE),obs = sample(1:100,100,replace = TRUE))

}
