% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict the new data set}
\usage{
predict(object, newX, ...)
}
\arguments{
\item{object}{the trained model that predictModel/predictModel_parallel function created}

\item{newX}{the new data that need to predict. It should have the same structure with trainset}

\item{...}{other parameters that belong to 'SuperLearner' package}
}
\value{
a list with predictions for both base model and meta model
}
\description{
Predicting the new data set using the trained ensemble models
}
