% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictModel.R
\name{predictModel}
\alias{predictModel}
\title{Prediction with single model or ensemble models with multi machine learning models}
\usage{
predictModel(Y, X, newX, base_model, cv = 10, method = "method.NNLS", ...)
}
\arguments{
\item{Y}{the independence, like certain environmental pollutants(eg. PM2.5) in training data that we
want to predict}

\item{X}{the other predictors(eg. temp, wind speed) in training data that we
want to use to predict air pollutants}

\item{newX}{the new test X data set that we use to predict}

\item{base_model}{character string in format of "SL.xgboost","SL.randomForest"}

\item{cv}{the number of cross validation}

\item{method}{the method of ensemble the base models, "method.NNLS" (the default).
Using "method.average" to set the weight as the average of all models.
more details see Superlearner}

\item{...}{other paramaters that belong to 'SuperLearner' pacakge}
}
\value{
a list including
ensemble_pre (the ensemble model predict value);
single_pre(the single model predict values);
the training assessment(R-squire and RMSE);
the basic model information;
}
\description{
Build the ensemble models and make prediciton using new dataset
}
\examples{
\dontrun{
#Let's predict the PM2.5 values with machine learming methods
#and an ensemble models automatically choosed based on the choosed models

#Spliting training set into two parts based on outcome: 80\% and 20\%
set.seed(1234)
data(envir_example)
index <-  caret::createDataPartition(envir_example$PM2.5, p=0.8, list=FALSE)
trainSet <- envir_example[ index,]
testSet <- envir_example[-index,]

#set the predictor "PM2.5" and dependence variables
predictor <- colnames(envir_example)[-c(1,6)]
dependence <- c("PM2.5")

##ensemble 6 kinds of multi models
sl_lib <- c("SL.xgboost","SL.randomForest","SL.ranger")
#get the predict model
pred_result <- predictModel(Y = trainSet[,dependence],X = trainSet[,predictor],
newX = testSet[,predictor], base_model = sl_lib)
# get the ensemble model result
head(pred_result$ensemble_pre)
# get the single model results
head(pred_result$single_pre)
}
}
