% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_ensemble.fit.R
\name{stack_ensemble.fit}
\alias{stack_ensemble.fit}
\title{Double-stack ensemble models with multi machine learning models}
\usage{
stack_ensemble.fit(
  Y,
  X,
  newX,
  base_model = c("SL.xgboost", "SL.ranger"),
  meta_model = c("SL.ranger"),
  original_feature = FALSE,
  ...
)
}
\arguments{
\item{Y}{the independence, like certain environmental pollutants(eg. PM2.5) in training data that we
want to predict}

\item{X}{the other predictors(eg. temp, wind speed) in training data that we
want to use to predict air pollutants}

\item{newX}{the new test X data set that we use to predict}

\item{base_model}{character string in format of "SL.xgboost","SL.randomForest"}

\item{meta_model}{character string in format of "SL.xgboost","SL.randomForest"
as the stack ensemble models}

\item{original_feature}{a Boolean value, whether integrate the original feature
into the meta model to build the stack ensemble models.}

\item{...}{other paramaters that belong to 'SuperLearner' pacakge}
}
\value{
a list with components:
the information about base model;
the information about meta model;
ensemble_pre (the ensemble model predict value);
stack_ensemble_pred (the double stack ensemble model predict value);
the double stack ensemble model assessment(R-squire and RMSE)
}
\description{
Double-stack ensemble models with multi machine learning models
}
