% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneModel.R
\name{tuningModel}
\alias{tuningModel}
\title{Tuning the parameters for the best or meta model}
\usage{
tuningModel(basemodel = "SL.ranger", params = list(), tune = list(), ...)
}
\arguments{
\item{basemodel}{Character string of the learner function}

\item{params}{List with parameters}

\item{tune}{List of hyperparameter settings}

\item{...}{other parameters, see SuperLearner}
}
\value{
a function with the model parameters
}
\description{
Tuning the parameters of a single model
}
\examples{
\dontrun{
set.seed(1234)
data(envir_example)
index <- caret::createDataPartition(envir_example$PM2.5, p = 0.8, list = FALSE)
trainSet <- envir_example[index, ]
testSet <- envir_example[-index, ]

# set the predictor "PM2.5" and dependence variables
predictor <- colnames(envir_example)[-c(1, 6)]
dependence <- c("PM2.5")

# create a new model with 3 types of 'mtry'(There are 3 models created)
learner <- tuningModel("SL.ranger",
  params = list(num.trees = 100),
  tune = list(mtry = c(1, 3, 7))
)

## ensemble created models
predictModel(
  Y = trainSet[, dependence], X = trainSet[, predictor],
  newX = testSet[, predictor], base_model = learner
)
}
}
