% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_ensemble.R
\name{stack_ensemble}
\alias{stack_ensemble}
\title{DEML framework with multiple meta machine learning models}
\usage{
stack_ensemble(
  object,
  meta_model = c("SL.ranger"),
  original_feature = FALSE,
  X,
  ...
)
}
\arguments{
\item{object}{the object that predictModel function created}

\item{meta_model}{character string in format of "SL.xgboost","SL.randomForest"
as the stack ensemble models}

\item{original_feature}{a Boolean value, whether integrate the original feature
into the meta model to build the stack ensemble models.}

\item{X}{optional when 'original_feature = FALSE', the predictors
(eg. temp, wind speed) in training data that we}

\item{...}{other paramaters that belong to 'SuperLearner' pacakge}
}
\value{
a list with components:
the information about base model;
the information about meta model;
ensemble_pre (the ensemble model predict value) when object has the newX parameter;
stack_ensemble_pred (the double stack ensemble model predict value) when object has the newX parameter;
the DEML model assessment(R-squire and RMSE)
}
\description{
Stacked ensemble models with multiple meta machine learning models
}
\examples{
\dontrun{
# Let's predict the PM2.5 values with machine learming methods

# Splitting training set into two parts based on outcome: 80\% and 20\%
set.seed(1234)
data("envir_example")
size <-
  caret::createDataPartition(y = envir_example$PM2.5, p = 0.8, list = FALSE)
trainset <- envir_example[size, ]
testset <- envir_example[-size, ]

# set the predictor "PM2.5" and dependence variables
y <- c("PM2.5")
x <- colnames(envir_example[-c(1, 6)]) # except "date" and "PM2.5"

# object do not include newX dataset
pre_stack_2 <-
  stack_ensemble(
    object = pred_m2,
    meta_model = c("SL.ranger", "SL.xgboost", "SL.lm"),
    original_feature = FALSE,
    X = testset[, x]
  )

# prediction
pred_stack_2_new <- predict(object = pre_stack_2, newX = testset[, x])
}
}
